backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
pydantic==2.9.2
pydantic-settings==2.6.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
